feat: start on backend implementation
Started on the implementation for receiving workouts from an api call to the server.
However I did not manage to finish this as I have problems with the "white screen"-problem I've had before.
Currently I get a response from the server, but can't manage to implement that so it shows on the screen
Refs: #27